Methods list

MAT.matopen

MAT.matread

MAT.matwrite

Methods usage

MAT.matopen — Function

matopen(filename [, mode]; compress = false) -> handle
matopen(f::Function, filename [, mode]; compress = false) -> f(handle)

Mode defaults to "r" for read. It can also be "w" for write, or "r+" for read or write without creation or truncation.

Compression on reading is detected/handled automatically; the compress keyword argument only affects write operations.

Use with read, write, close, keys, and haskey.

MAT.matread — Method

matread(filename) -> Dict

Return a dictionary of all the variables and values in a Matlab file, opening and closing it automatically.

MAT.matwrite — Method

matwrite(filename, d::Dict; compress::Bool = false, version::String)

Write a dictionary containing variable names as keys and values as values to a Matlab file, opening and closing it automatically.